% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_server.R
\name{set_var}
\alias{set_var}
\title{#' @export
prism_get_output_structure<-function()
{
out<-list(
n_agents=prism_output(source="$n_agents", type = "numeric/scalar", group = "", title = "Number of simulated individuals", description = ""),
)
return(out)
}}
\usage{
set_var(variable, value)
}
\description{
#' @export
prism_get_output_structure<-function()
{
out<-list(
n_agents=prism_output(source="$n_agents", type = "numeric/scalar", group = "", title = "Number of simulated individuals", description = ""),
)
return(out)
}
}
